requirements.txt
streamlit
python-dotenv
agno
pandas
groq
googlesearch-python
pycountry